+++ webknossos/webknossos/cli/convert_knossos.py
@@ -307,7 +307,7 @@ def main(
         Mag,
         typer.Option(
             help="Mag to start upsampling from. "
-            "Should be number or minus separated string (e.g. 2 or 2-2-2).",
+            "Should be number or hyphen-separated string (e.g. 2 or 2-2-2).",
             parser=parse_mag,
             metavar="MAG",
         ),